Notice the mock environment return float64(double) type.

+++ fagents/tools/mock_environment.py
@@ -33,13 +33,13 @@ class MockEnvironment(object):
   def observation_space(self):
     low = np.zeros(self._observ_shape)
     high = np.ones(self._observ_shape)
-    return gym.spaces.Box(low, high)
+    return gym.spaces.Box(low, high, dtype=np.float32)
 
   @property
   def action_space(self):
     low = np.zeros(self._action_shape)
     high = np.ones(self._action_shape)
-    return gym.spaces.Box(low, high)
+    return gym.spaces.Box(low, high, dtype=np.float32)
 
   @property
   def unwrapped(self):
@@ -62,7 +62,7 @@ class MockEnvironment(object):
     return self._current_observation()
 
   def _current_observation(self):
-    return self._random.uniform(0, 1, self._observ_shape)
+    return self._random.uniform(0, 1, self._observ_shape).astype(np.float32)
 
   def _current_reward(self):
-    return self._random.uniform(-1, 1)
+    return self._random.uniform(-1, 1).astype(np.float32)
